refactor: UI components and styles have been reorganized
