+++ src/styles/components/UI/card.scss
@@ -1,4 +1,4 @@
-@import '../variables.scss';
+@import '../../variables.scss';
 
 .card-wrapper {
   padding: 0 6px;
