+++ src/styles/components/UI/loader.scss
@@ -1,4 +1,4 @@
-@import '../variables.scss';
+@import '../../variables.scss';
 
 .loader {
   width: 6rem;
