+++ src/styles/components/UI/modal.scss
@@ -1,4 +1,4 @@
-@import '../variables.scss';
+@import '../../variables.scss';
 
 .modal {
   display: none;
