+++ src/styles/components/UI/noAvailable.scss
