+++ src/styles/globals.scss
@@ -2,7 +2,8 @@
 
 #root {
   height: 100%;
-  min-height: 850px;
+  min-height: 100vh;
+  padding-bottom: 100px;
   background: linear-gradient(0deg, $blue-100 20%, $blue-900 100%) no-repeat;
 }
