+++ src/styles/views/movies.scss
@@ -1,0 +1,11 @@
+.movies {
+  overflow-y: auto;
+  padding-top: 100px;
+  padding-left: 50px;
+  padding-right: 50px;
+  height: 100vh;
+  position: absolute;
+  top: 0;
+  left: 0;
+  width: 100%;
+}
